lojas_giranianas/requirements.txt
http://www.djangoproject.com/download/1.3-beta-1/tarball/
MySQL-Python
south
django-registration
